app/requirements.txt
beautifulsoup4==4.4.1
Django==1.8
django-app-namespace-template-loader==0.3
django-blog-zinnia==0.15.2
django-contrib-comments==1.6.2
django-mptt==0.8.0
django-photologue==3.5.1
django-sortedm2m==1.1.1
django-tagging==0.4.1
django-xmlrpc==0.1.5
ExifRead==2.1.2
mots-vides==2015.5.11
Pillow==3.1.0
pyparsing==2.0.7
pytz==2015.7
six==1.10.0
zinnia-theme-bootstrap==0.4
gunicorn==19.4.5
psycopg2==2.6.1
